requirements.txt
streamlit
python-docx
docxtpl
pillow
